extract CORS middleware policy into SRP core crate
