+++ Cargo.toml
@@ -57,6 +57,7 @@ members = [
   "crates/bitnet-eval-core",     # SRP: deterministic eval/scoring math helpers
   "crates/bitnet-feature-matrix",
   "crates/bitnet-client-ip-core", # SRP: reusable client IP extraction from forwarding headers
+  "crates/bitnet-cors-policy-core", # SRP: reusable CORS policy primitives for middleware stacks
   "crates/bitnet-api-versioning-core", # SRP: reusable API version negotiation and routing helpers
   "crates/bitnet-api-key-auth-core", # SRP: API key auth policy/store primitives
   "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
