+++ crates/bitnet-cors-policy-core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "bitnet-cors-policy-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "SRP core CORS policy primitives for server middleware configuration"
+keywords.workspace = true
+categories.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[lints]
+workspace = true
